requirements.txt
asgiref==3.7.2
Django==4.2.3
openai==0.27.8
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
# Optional/runtime
gunicorn>=20.1.0
psycopg2-binary>=2.9
